+++ src/symmetric_crypto/mod.rs
@@ -9,6 +9,8 @@ mod aes_256_gcm;
 mod chacha20_poly1305;
 mod dem;
 mod key;
+#[cfg(feature = "rfc5649")]
+mod key_wrapping_rfc_5649;
 mod nonce;
 #[cfg(feature = "chacha")]
 mod xchacha20_poly1305;
@@ -21,6 +23,8 @@ pub use aes_256_gcm::Aes256Gcm;
 pub use chacha20_poly1305::ChaCha20Poly1305;
 pub use dem::{Dem, DemInPlace, DemStream, Instantiable};
 pub use key::SymmetricKey;
+#[cfg(feature = "rfc5649")]
+pub use key_wrapping_rfc_5649::{unwrap, unwrap_64, wrap, wrap_64};
 pub use nonce::Nonce;
 #[cfg(feature = "chacha")]
 pub use xchacha20_poly1305::XChaCha20Poly1305;